Fix Cache invalidation bug
Replace the Entity Set by a Map in Cache.java in order to fix the bug
where the Set keeps getting bigger at each Model creation, and never
cleaned, thus resulting in Entities being in an invalid state.
Also added a getCacheID method in Model.java to be used as the Map key.