Add AA_DB_RESET flag
There are certain SQLite features that are hard to
set or modify in an existing DB, such as:

 - Modify foreign key actions
 - Rename columns or tables
 - `ON (UPDATE|CREATE)` actions

They usually require creating a new table, dropping
the old one, and renaming the new one. If a full
db reset is performed then these definitions can be
included without migration scripts.

Reset capabilities are added with these changes. A
db delete/create will be issued whenever the
`AA_DB_VERSION` changes and the following lines
exist in the app's manifest:

```
<meta-data
    android:name="AA_DB_RESET"
    android:value="true" />
```

The reset will be checked in the `Cache.initialize`
method.